pyproject.toml
[tool.poetry]
name = "gocqapi"
version = "0.1.6"
description = "go-cqhttp API typing annoations, return data models and utils for nonebot"
authors = ["风屿 <i@windis.cn>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7.3, <4.0.0"
nonebot2 = "^2.0.0-beta.5"
nonebot-adapter-onebot = "^2.1.3"
httpx = "^0.23.0"
pydantic = "^1.9.0"


[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = {version = "^21.11b1", allow-prereleases = true}
isort = "^5.10.1"

[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"
